requirements.txt
pandas
streamlit
snowflake-snowpark-python
